clean up locking; prefix lockfile with "."
