+++ lib/install.js
@@ -73,7 +73,6 @@ var npm = require("./npm.js")
   , log = require("npmlog")
   , path = require("path")
   , fs = require("graceful-fs")
-  , crypto = require("crypto")
   , writeFileAtomic = require("write-file-atomic")
   , cache = require("./cache.js")
   , asyncMap = require("slide").asyncMap
@@ -86,7 +85,9 @@ var npm = require("./npm.js")
   , sortedObject = require("sorted-object")
   , mapToRegistry = require("./utils/map-to-registry.js")
   , npa = require("npm-package-arg")
-  , lockfile = require("lockfile")
+  , locker = require("./utils/locker.js")
+  , lock = locker.lock
+  , unlock = locker.unlock
 
 function install (args, cb_) {
   var hasArguments = !!args.length
@@ -1048,46 +1049,4 @@ function prepareForInstallMany (packageData, depsKey, bundled, wrap, family) {
     t = d + "@" + t
     return t
   })
-}
-
-// copied from lib/util/locker.js because I don't feel like dealing with
-// that module's desire to run getCacheStat
-var installLocks = {}
-function lockFileName (base, name) {
-  var c = name.replace(/[^a-zA-Z0-9]+/g, "-").replace(/^-+|-+$/g, "")
-    , h = crypto.createHash("sha1").update(name).digest("hex")
-  var generated = h.substr(0, 8) + "-" + c.substr(-32)
-  log.silly("lockFile", generated, name)
-  return path.resolve(base, generated + ".lock")
-}
-
-function lock (base, name, cb) {
-  mkdir(base, function (er) {
-    if (er) return cb(er)
-
-    var opts = { stale:   npm.config.get("cache-lock-stale")
-               , retries: npm.config.get("cache-lock-retries")
-               , wait:    npm.config.get("cache-lock-wait") }
-    var lf = lockFileName(base, name)
-    log.verbose("lock", name, lf)
-    lockfile.lock(lf, opts, function(er) {
-      if (!er) installLocks[lf] = true
-      cb(er)
-    })
-  })
-}
-
-function unlock (base, name, cb) {
-  var lf = lockFileName(base, name)
-    , locked = installLocks[lf]
-  if (locked === false) {
-    return process.nextTick(cb)
-  }
-  else if (locked === true) {
-    installLocks[lf] = false
-    lockfile.unlock(lf, cb)
-  }
-  else {
-    throw new Error("Attempt to unlock " + name + ", which hasn't been locked")
-  }
 }
