+++ lib/utils/locker.js
@@ -2,51 +2,51 @@ var crypto = require("crypto")
 var path = require("path")
 
 var npm = require("../npm.js")
-var lockFile = require("lockfile")
+var lockfile = require("lockfile")
 var log = require("npmlog")
-var getCacheStat = require("../cache/get-stat.js")
+var mkdirp = require("mkdirp")
 
-function lockFileName (u) {
-  var c = u.replace(/[^a-zA-Z0-9]+/g, "-").replace(/^-+|-+$/g, "")
-    , h = crypto.createHash("sha1").update(u).digest("hex")
-  h = h.substr(0, 8)
-  c = c.substr(-32)
-  log.silly("lockFile", h + "-" + c, u)
-  return path.resolve(npm.config.get("cache"), h + "-" + c + ".lock")
+var installLocks = {}
+function lockFileName (base, name) {
+  var c = name.replace(/[^a-zA-Z0-9]+/g, "-").replace(/^-+|-+$/g, "")
+    , h = crypto.createHash("sha1").update(name).digest("hex")
+  var generated = "." + h.substr(0, 8) + "-" + c.substr(-32)
+  log.silly("lockFile", generated, name)
+  return path.resolve(base, generated + ".lock")
 }
 
-var myLocks = {}
-function lock (u, cb) {
-  // the cache dir needs to exist already for this.
-  getCacheStat(function (er) {
+function lock (base, name, cb) {
+  mkdirp(base, function (er) {
     if (er) return cb(er)
-    var opts = { stale: npm.config.get("cache-lock-stale")
+
+    var opts = { stale:   npm.config.get("cache-lock-stale")
                , retries: npm.config.get("cache-lock-retries")
-               , wait: npm.config.get("cache-lock-wait") }
-    var lf = lockFileName(u)
-    log.verbose("lock", u, lf)
-    lockFile.lock(lf, opts, function(er) {
-      if (!er) myLocks[lf] = true
+               , wait:    npm.config.get("cache-lock-wait") }
+    var lf = lockFileName(base, name)
+    log.verbose("lock", name, lf)
+    lockfile.lock(lf, opts, function(er) {
+      if (!er) installLocks[lf] = true
       cb(er)
     })
   })
 }
 
-function unlock (u, cb) {
-  var lf = lockFileName(u)
-    , locked = myLocks[lf]
+function unlock (base, name, cb) {
+  var lf = lockFileName(base, name)
+    , locked = installLocks[lf]
   if (locked === false) {
     return process.nextTick(cb)
-  } else if (locked === true) {
-    myLocks[lf] = false
-    lockFile.unlock(lockFileName(u), cb)
-  } else {
-    throw new Error("Attempt to unlock " + u + ", which hasn't been locked")
+  }
+  else if (locked === true) {
+    installLocks[lf] = false
+    lockfile.unlock(lf, cb)
+  }
+  else {
+    throw new Error("Attempt to unlock " + name + ", which hasn't been locked")
   }
 }
 
 module.exports = {
-  lock: lock,
-  unlock: unlock,
-  _lockFileName: lockFileName
+  lock   : lock,
+  unlock : unlock
 }
